Move setupControllers() into ::Actions, call from ::WordPressActionsAndFilters
